Some fixes to make it compile with Visual Studio 2015.

+++ DynamicAudioNormalizerVST/src/Spooky.cpp
@@ -257,7 +257,7 @@ void hash128(const void *const message, size_t length, uint64_t *hash1, uint64_t
 	remainder = (length - ((const uint8_t*)end - (const uint8_t*)message));
 	memcpy(buf, end, remainder);
 	memset(((uint8_t*)buf) + remainder, 0, sc_blockSize - remainder);
-	((uint8_t*)buf)[sc_blockSize - 1] = ((uint32_t)remainder);
+	((uint8_t*)buf)[sc_blockSize - 1] = (uint8_t)((uint32_t)remainder);
 
 	// do some final mixing 
 	End(buf, h0, h1, h2, h3, h4, h5, h6, h7, h8, h9, h10, h11);
